pyproject.toml
[project]
name = "pymuse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.21.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.16.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.41",
    "yt-dlp>=2025.6.30",
]
